+++ Letos/gui/uiconfig.h
@@ -177,7 +177,6 @@ CFG_UI_CATEGORIES(Ui,
         CFG_ENTRY(bool,                    ShowDataViewTooltips,        true)
         CFG_ENTRY(bool,                    KeepNullWhenEmptyValue,      true)
         CFG_ENTRY(bool,                    UseSciFormatForDoubles,      false)
-        CFG_ENTRY(bool,                    UseLfForMultilineEditors,    false)
         CFG_ENTRY(Cfg::ColorPickerConfig,  ColorPickerConfig,           Cfg::ColorPickerConfig())
         CFG_ENTRY(int,                     ToolBarIconSize,             0)
         CFG_ENTRY(QString,                 LastUsedConfigPage,          QString())